+++ fltk-sys/src/browser.rs
@@ -348,9 +348,6 @@ extern "C" {
     pub fn Fl_Browser_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
-    pub fn Fl_Browser_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
     pub fn Fl_Browser_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
@@ -882,9 +879,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Hold_Browser_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn Fl_Hold_Browser_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
     pub fn Fl_Hold_Browser_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
@@ -1464,9 +1458,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Select_Browser_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn Fl_Select_Browser_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
     pub fn Fl_Select_Browser_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
@@ -2057,9 +2048,6 @@ extern "C" {
     pub fn Fl_Multi_Browser_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
-    pub fn Fl_Multi_Browser_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
     pub fn Fl_Multi_Browser_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
@@ -2621,9 +2609,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_File_Browser_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn Fl_File_Browser_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
     pub fn Fl_File_Browser_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
@@ -3205,9 +3190,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Check_Browser_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn Fl_Check_Browser_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
     pub fn Fl_Check_Browser_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);