+++ fltk-sys/src/frame.rs
@@ -348,8 +348,5 @@ extern "C" {
     pub fn Fl_Box_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
-    pub fn Fl_Box_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
     pub fn Fl_Box_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }