+++ fltk-sys/src/table.rs
@@ -348,9 +348,6 @@ extern "C" {
     pub fn Fl_Table_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
-    pub fn Fl_Table_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
     pub fn Fl_Table_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
@@ -997,9 +994,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Table_Row_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn Fl_Table_Row_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
     pub fn Fl_Table_Row_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);