+++ fltk-sys/src/text.rs
@@ -668,9 +668,6 @@ extern "C" {
     pub fn Fl_Text_Display_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
-    pub fn Fl_Text_Display_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
     pub fn Fl_Text_Display_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
@@ -1328,9 +1325,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Text_Editor_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn Fl_Text_Editor_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
     pub fn Fl_Text_Editor_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
@@ -2175,9 +2169,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Simple_Terminal_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn Fl_Simple_Terminal_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
     pub fn Fl_Simple_Terminal_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);