+++ fltk-sys/src/tree.rs
@@ -348,9 +348,6 @@ extern "C" {
     pub fn Fl_Tree_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
-    pub fn Fl_Tree_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
     pub fn Fl_Tree_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 #[repr(C)]