+++ fltk-sys/src/widget.rs
@@ -343,8 +343,5 @@ extern "C" {
     pub fn Fl_Widget_super_draw_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }
 extern "C" {
-    pub fn Fl_Widget_super_handle(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
-}
-extern "C" {
     pub fn Fl_Widget_super_handle_first(ptr: *mut Fl_Widget, flag: ::std::os::raw::c_int);
 }